hw1/requirements.txt
annotated-types==0.7.0
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
filelock==3.19.1
flake8==7.3.0
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.46
idna==3.11
Jinja2==3.1.6
joblib==1.5.2
MarkupSafe==2.1.5
mccabe==0.7.0
mpmath==1.3.0
networkx==3.5
numpy==2.3.3
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.21.5
nvidia-nvtx-cu11==11.8.86
packaging==26.0
pillow==11.3.0
platformdirs==4.9.4
protobuf==6.33.6
pycodestyle==2.14.0
pydantic==2.12.5
pydantic_core==2.41.5
pyflakes==3.4.0
PyYAML==6.0.3
requests==2.32.5
scikit-learn==1.7.2
scipy==1.16.3
sentry-sdk==2.55.0
setuptools==80.9.0
smmap==5.0.3
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.7.1+cu118
torchaudio==2.7.1+cu118
torchvision==0.22.1+cu118
tqdm==4.67.1
triton==3.3.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
wandb==0.25.1
wheel==0.45.1
